@@ -74,7 +74,7 @@
     "postcss-loader": "^0.8.2",
     "precss": "^1.4.0",
     "react-addons-test-utils": "^15.2.1",
-    "react-hot-loader": "^1.3.0",
+    "react-hot-loader": "3.0.0-beta.2",
     "redux-devtools": "^3.2.0",
     "redux-devtools-dock-monitor": "^1.1.1",
     "redux-devtools-log-monitor": "^1.0.9",
@@ -93,24 +93,32 @@
     "babel-polyfill": "^6.7.4",
     "bootstrap": "^3.3.6",
     "classnames": "^2.2.3",
+    "create-react-class": "^15.5.2",
     "font-awesome": "^4.5.0",
     "graphql-tag": "^0.1.15",
     "jquery": "^2.2.1",
     "js-base64": "^2.1.9",
+    "lodash": "^4.17.4",
     "moment": "^2.14.1",
-    "react": "^15.3.0",
+    "prop-types": "^15.5.8",
+    "react": "^15.5.0",
     "react-addons-shallow-compare": "^15.3.0",
     "react-apollo": "^0.5.13",
-    "react-bootstrap": "^0.28.3",
+    "react-bootstrap": "^0.28.5",
     "react-dom": "^15.3.0",
+    "react-intl": "^2.2.3",
     "react-modal": "^0.6.1",
     "react-motion": "^0.4.4",
     "react-notification": "^5.0.1",
+    "react-radio-group": "^3.0.1",
     "react-redux": "^4.4.1",
     "react-router": "^2.0.0",
     "react-router-redux": "^4.0.1",
+    "react-select": "^1.0.0-rc.3",
+    "react-select-box": "^3.0.1",
     "react-tap-event-plugin": "^2.0.1",
     "redux": "^3.3.1",
+    "redux-form": "^6.6.2",
     "redux-thunk": "^2.0.1",
     "whatwg-fetch": "^1.0.0"
   }
