@@ -5,10 +5,7 @@ const dateFormat = 'DD/MM/YYYY HH:mm';
   constants
  ------------------------------------------*/
 // non protected views:
-const ENTER_HOME_VIEW  = 'ENTER_HOME_VIEW';
-const LEAVE_HOME_VIEW  = 'LEAVE_HOME_VIEW';
-const ENTER_ABOUT_VIEW = 'ENTER_ABOUT_VIEW';
-const LEAVE_ABOUT_VIEW = 'LEAVE_ABOUT_VIEW';
+const REQUEST_SUBMIT = 'REQUEST_SUBMIT';
 const ENTER_LOGIN_VIEW = 'ENTER_LOGIN_VIEW';
 const LEAVE_LOGIN_VIEW = 'LEAVE_LOGIN_VIEW';
 const ENTER_REGISTER_VIEW = 'ENTER_REGISTER_VIEW';
@@ -33,8 +30,6 @@ export default function (state = initialState, action) {
   // /////////////////////
   // non protected views:
   // /////////////////////
-  case ENTER_HOME_VIEW:
-  case ENTER_ABOUT_VIEW:
   case ENTER_LOGIN_VIEW:
   case ENTER_REGISTER_VIEW:
     // can't enter if you are already inside
@@ -46,8 +41,6 @@ export default function (state = initialState, action) {
       };
     }
     return state;
-  case LEAVE_HOME_VIEW:
-  case LEAVE_ABOUT_VIEW:
   case LEAVE_LOGIN_VIEW:
   case LEAVE_REGISTER_VIEW:
     // can't leave if you aren't already inside
@@ -90,30 +83,10 @@ export default function (state = initialState, action) {
 /* -----------------------------------------
   Reducer
  ------------------------------------------*/
-export function enterHome() {
+export function onSubmit() {
   return {
-    type:         ENTER_HOME_VIEW,
-    currentView:  'home'
-  };
-}
-export function leaveHome() {
-  return {
-    type:         LEAVE_HOME_VIEW,
-    currentView:  'home'
-  };
-}
-
-export function enterAbout() {
-  return {
-    type:         ENTER_ABOUT_VIEW,
-    currentView:  'about'
-  };
-}
-export function leaveAbout() {
-  return {
-    type:         LEAVE_ABOUT_VIEW,
-    currentView:  'about'
-  };
+    type: REQUEST_SUBMIT,
+  }
 }
 
 export function enterLogin() {
